homework is corrected again updated_comments
