+++ homework/Daria_Smirnova/test_api_tripleaction/endpoints/base_endpoint.py
@@ -1,4 +1,5 @@
 import allure
+import requests
 
 
 class BaseEndpoint:
@@ -12,7 +13,7 @@ class BaseEndpoint:
         self.json = None
 
     @allure.step("Check response code is 200")
-    def check_status_code_is_correct(self):
+    def check_status_code_is_200(self):
         assert self.response.status_code == 200, "Status code is not 200"
 
     @allure.step("Check response color is the same as sent")
@@ -26,3 +27,10 @@ class BaseEndpoint:
     @allure.step("Check response name is correct")
     def check_name_correct(self, expected_name):
         assert self.json["name"] == expected_name, "Name is not correct"
+
+    @allure.step("Delete object")
+    def delete_object(self, object_id, headers=None):
+        headers = headers or self.headers
+        delete_url = f"{self.url}/{object_id}"
+        self.response = requests.delete(delete_url, headers=headers)
+        return self.response
